requirements.txt
easyocr
selenium
schedule
python-dotenv
torch
torchvision
torchaudio